@@ -8,7 +8,7 @@
 
 import UIKit
 
-class SocketEventObserver: NSObject, URLSessionWebSocketDelegate {
+actor SocketEventObserver: NSObject, URLSessionWebSocketDelegate {
 
     private let id = UUID().uuidString.suffix(4)
     private var continuation: CheckedContinuation<URLSessionWebSocketTask, Error>?
@@ -29,7 +29,44 @@ class SocketEventObserver: NSObject, URLSessionWebSocketDelegate {
 
     // MARK: - Helpers
 
-    func resume(url: URL) {
+    nonisolated func start(with url: URL) {
+        Task { await _start(with: url) }
+    }
+
+    nonisolated func invalidate() {
+        Task { await _invalidate() }
+    }
+
+    // MARK: - URLSessionWebSocketDelegate
+
+    nonisolated func urlSession(
+        _ session: URLSession,
+        webSocketTask: URLSessionWebSocketTask,
+        didOpenWithProtocol protocol: String?
+    ) {
+        Task { await _continue(
+            withWebSocketTask: webSocketTask,
+            result: .success(webSocketTask)
+        )}
+    }
+
+    nonisolated func urlSession(
+        _ session: URLSession,
+        webSocketTask: URLSessionWebSocketTask,
+        didCloseWith closeCode: URLSessionWebSocketTask.CloseCode,
+        reason: Data?
+    ) {
+        Task { await _continue(
+            withWebSocketTask: webSocketTask,
+            result: .failure(WebSocketError.didClose)
+        )}
+    }
+
+    // MARK: - Helpers
+
+    private func _start(
+        with url: URL
+    ) {
         let operationQueue = OperationQueue()
         operationQueue.maxConcurrentOperationCount = 1
 
@@ -38,42 +75,26 @@ class SocketEventObserver: NSObject, URLSessionWebSocketDelegate {
             delegate: self,
             delegateQueue: operationQueue
         )
+
         self.session = session
 
         socket = session.webSocketTask(with: url)
         socket?.resume()
     }
 
-    func invalidate() {
+    private func _continue(
+        withWebSocketTask webSocketTask: URLSessionWebSocketTask,
+        result: Result<URLSessionWebSocketTask, Error>
+    ) {
+        guard !isSetup else { return }
+        isSetup = true
+        assert(socket === webSocketTask)
+        socket = nil
+        continuation?.resume(with: result)
+        continuation = nil
+    }
+
+    private func _invalidate() {
         session?.invalidateAndCancel()
-    }
-
-    // MARK: - URLSessionWebSocketDelegate
-
-    func urlSession(
-        _ session: URLSession,
-        webSocketTask: URLSessionWebSocketTask,
-        didOpenWithProtocol protocol: String?
-    ) {
-        guard !isSetup else { return }
-        isSetup = true
-        assert(socket === webSocketTask)
-        socket = nil
-        continuation?.resume(returning: webSocketTask)
-        continuation = nil
-    }
-
-    func urlSession(
-        _ session: URLSession,
-        webSocketTask: URLSessionWebSocketTask,
-        didCloseWith closeCode: URLSessionWebSocketTask.CloseCode,
-        reason: Data?
-    ) {
-        guard !isSetup else { return }
-        isSetup = true
-        assert(socket === webSocketTask)
-        socket = nil
-        continuation?.resume(throwing: WebSocketError.didClose)
-        continuation = nil
     }
 }
